+++ clients/algoliasearch-client-ruby/lib/algolia/api/search_client.rb
@@ -34,6 +34,18 @@ module Algolia
     end
 
     def self.create_with_config(config)
+      if config.connect_timeout.nil? || config.connect_timeout.empty?
+        config.connect_timeout = 2000
+      end
+
+      if config.read_timeout.nil? || config.read_timeout.empty?
+        config.read_timeout = 5000
+      end
+
+      if config.write_timeout.nil? || config.write_timeout.empty?
+        config.write_timeout = 30000
+      end
+
       new(config)
     end
 
@@ -207,6 +219,7 @@ module Algolia
     #
     # Required API Key ACLs:
     #   - admin
+    # THIS OPERATION IS DEPRECATED
     # @param x_algolia_user_id [String] Unique identifier of the user who makes the search request. (required)
     # @param assign_user_id_params [AssignUserIdParams]  (required)
     # @param request_options: The request options to send along with the query, they will be merged with the transporter base parameters (headers, query params, timeouts, etc.). (optional)
@@ -304,6 +317,7 @@ module Algolia
     #
     # Required API Key ACLs:
     #   - admin
+    # THIS OPERATION IS DEPRECATED
     # @param x_algolia_user_id [String] Unique identifier of the user who makes the search request. (required)
     # @param batch_assign_user_ids_params [BatchAssignUserIdsParams]  (required)
     # @param request_options: The request options to send along with the query, they will be merged with the transporter base parameters (headers, query params, timeouts, etc.). (optional)
@@ -1673,6 +1687,7 @@ module Algolia
     #
     # Required API Key ACLs:
     #   - admin
+    # THIS OPERATION IS DEPRECATED
     # @param request_options: The request options to send along with the query, they will be merged with the transporter base parameters (headers, query params, timeouts, etc.). (optional)
     # @return [Http::Response] the response
     def get_top_user_ids_with_http_info(request_options = {})
@@ -1710,6 +1725,7 @@ module Algolia
     #
     # Required API Key ACLs:
     #   - admin
+    # THIS OPERATION IS DEPRECATED
     # @param user_id [String] Unique identifier of the user who makes the search request. (required)
     # @param request_options: The request options to send along with the query, they will be merged with the transporter base parameters (headers, query params, timeouts, etc.). (optional)
     # @return [Http::Response] the response
@@ -1754,6 +1770,7 @@ module Algolia
     #
     # Required API Key ACLs:
     #   - admin
+    # THIS OPERATION IS DEPRECATED
     # @param get_clusters [Boolean] Whether to include the cluster's pending mapping state in the response.
     # @param request_options: The request options to send along with the query, they will be merged with the transporter base parameters (headers, query params, timeouts, etc.). (optional)
     # @return [Http::Response] the response
@@ -1834,6 +1851,7 @@ module Algolia
     #
     # Required API Key ACLs:
     #   - admin
+    # THIS OPERATION IS DEPRECATED
     # @param request_options: The request options to send along with the query, they will be merged with the transporter base parameters (headers, query params, timeouts, etc.). (optional)
     # @return [Http::Response] the response
     def list_clusters_with_http_info(request_options = {})
@@ -1914,6 +1932,7 @@ module Algolia
     #
     # Required API Key ACLs:
     #   - admin
+    # THIS OPERATION IS DEPRECATED
     # @param page [Integer] Requested page of the API response. If `null`, the API response is not paginated.
     # @param hits_per_page [Integer] Number of hits per page. (default to 100)
     # @param request_options: The request options to send along with the query, they will be merged with the transporter base parameters (headers, query params, timeouts, etc.). (optional)
@@ -2129,6 +2148,7 @@ module Algolia
     #
     # Required API Key ACLs:
     #   - admin
+    # THIS OPERATION IS DEPRECATED
     # @param user_id [String] Unique identifier of the user who makes the search request. (required)
     # @param request_options: The request options to send along with the query, they will be merged with the transporter base parameters (headers, query params, timeouts, etc.). (optional)
     # @return [Http::Response] the response
@@ -2906,6 +2926,7 @@ module Algolia
     #
     # Required API Key ACLs:
     #   - admin
+    # THIS OPERATION IS DEPRECATED
     # @param search_user_ids_params [SearchUserIdsParams]  (required)
     # @param request_options: The request options to send along with the query, they will be merged with the transporter base parameters (headers, query params, timeouts, etc.). (optional)
     # @return [Http::Response] the response