+++ docs/snippets/java/src/test/java/com/algolia/Ingestion.java
@@ -106,6 +106,7 @@ class SnippetIngestionClient {
             .setLocales(Arrays.asList("de"))
             .setUrl("http://commercetools.com")
             .setProjectKey("keyID")
+            .setProductQueryPredicate("masterVariant(attributes(name=\"Brand\" and value=\"Algolia\"))")
         )
         .setAuthenticationID("6c02aeb1-775e-418e-870b-1faccd4b2c0f")
     );