+++ tests/output/javascript/src/client/abtesting.test.ts
@@ -22,15 +22,15 @@ describe('commonApi', () => {
     expect(decodeURIComponent(result.algoliaAgent)).toMatch(
       /^Algolia for JavaScript \(\d+\.\d+\.\d+(-?.*)?\)(; [a-zA-Z. ]+ (\(\d+((\.\d+)?\.\d+)?(-?.*)?\))?)*(; Abtesting (\(\d+\.\d+\.\d+(-?.*)?\)))(; [a-zA-Z. ]+ (\(\d+((\.\d+)?\.\d+)?(-?.*)?\))?)*$/,
     );
-  }, 15000);
+  }, 25000);
 
   test('the user agent contains the latest version', async () => {
     const client = createClient();
 
     const result = (await client.customPost({ path: '1/test' })) as unknown as EchoResponse;
 
-    expect(decodeURIComponent(result.algoliaAgent)).toMatch(/^Algolia for JavaScript \(5.21.0\).*/);
-  }, 15000);
+    expect(decodeURIComponent(result.algoliaAgent)).toMatch(/^Algolia for JavaScript \(5.23.1\).*/);
+  }, 25000);
 });
 
 describe('parameters', () => {
@@ -45,7 +45,7 @@ describe('parameters', () => {
     const result = (await client.getABTest({ id: 123 })) as unknown as EchoResponse;
 
     expect(result.host).toEqual('analytics.algolia.com');
-  }, 15000);
+  }, 25000);
 
   test('uses the correct region', async () => {
     const client = algoliasearch('my-app-id', 'my-api-key').initAbtesting({
@@ -58,7 +58,7 @@ describe('parameters', () => {
     const result = (await client.getABTest({ id: 123 })) as unknown as EchoResponse;
 
     expect(result.host).toEqual('analytics.us.algolia.com');
-  }, 15000);
+  }, 25000);
 
   test('throws when incorrect region is given', async () => {
     try {
@@ -74,7 +74,7 @@ describe('parameters', () => {
     } catch (e) {
       expect((e as Error).message).toMatch('`region` must be one of the following: de, us');
     }
-  }, 15000);
+  }, 25000);
 });
 
 describe('setClientApiKey', () => {
@@ -106,7 +106,7 @@ describe('setClientApiKey', () => {
 
       expect(result).toEqual({ headerAPIKeyValue: 'updated-api-key' });
     }
-  }, 15000);
+  }, 25000);
 });
 
 describe('init', () => {