+++ clients/algoliasearch-client-php/CHANGELOG.md
@@ -1,3 +1,15 @@
+## [4.18.1](https://github.com/algolia/algoliasearch-client-php/compare/4.18.0...4.18.1)
+
+- [a51683025e](https://github.com/algolia/api-clients-automation/commit/a51683025e) chore: deprecate mcm methods ([#4694](https://github.com/algolia/api-clients-automation/pull/4694)) by [@kai687](https://github.com/kai687/)
+
+## [4.18.0](https://github.com/algolia/algoliasearch-client-php/compare/4.17.0...4.18.0)
+
+- [c3addb449](https://github.com/algolia/api-clients-automation/commit/c3addb449) feat(specs): add CT Query predicate property ([#4654](https://github.com/algolia/api-clients-automation/pull/4654)) by [@morganleroi](https://github.com/morganleroi/)
+
+## [4.17.0](https://github.com/algolia/algoliasearch-client-php/compare/4.16.0...4.17.0)
+
+- [26457d0ec8](https://github.com/algolia/api-clients-automation/commit/26457d0ec8) feat(specs): add the owner attributes to ingestion resources ([#4616](https://github.com/algolia/api-clients-automation/pull/4616)) by [@millotp](https://github.com/millotp/)
+
 ## [4.16.0](https://github.com/algolia/algoliasearch-client-php/compare/4.15.1...4.16.0)
 
 - [ab623194e](https://github.com/algolia/api-clients-automation/commit/ab623194e) feat(specs): add `subscriptionAction` to ingestion specs ([#4587](https://github.com/algolia/api-clients-automation/pull/4587)) by [@shortcuts](https://github.com/shortcuts/)